Isolate core data transformation logic from db.
Core logic still requires knowledge of db model, but doesn't know
anything about how queries are executed or even connecting to a
database. All it knows is what method on the model needs called with the
transformed data as input in order to achieve the desired result.

Transaction route handlers now just defer to core methods to validate &
transform inputs, then calls the returned db model method with the
returned transformed data & sends the formats the result for Response.